requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
coverage==7.4.1
cryptography==41.0.7
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.1
MarkupSafe==2.1.4
packaging==23.2
passlib==1.7.4
pluggy==1.4.0
psycopg2==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pydub==0.25.1
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqladmin==0.16.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.26.0
WTForms==3.1.2

# Whisper Nemo dependencies 
Cython==3.0.7
nemo_toolkit[asr]==1.21.0
transformers>=4.26.1
wget==3.2
whisperx @ git+https://github.com/m-bain/whisperX.git@a5dca2cc65b1a37f32a347e574b2c56af3a7434a
demucs @ git+https://github.com/facebookresearch/demucs#egg=demucs
deepmultilingualpunctuation==1.0.1
torch==2.0.1
torchvision==0.15.2